--- tests/test_endtoend.py
@@ -1,396 +1,0 @@
-"""
-End-to-end Telepresence tests for running directly in the operating system.
-"""
-
-import os
-from sys import executable
-from json import (
-    JSONDecodeError,
-    loads, dumps,
-)
-from unittest import (
-    TestCase,
-)
-from subprocess import (
-    CalledProcessError,
-    PIPE, STDOUT, Popen, check_output, check_call,
-)
-from pathlib import Path
-from shutil import which
-
-from .utils import (
-    KUBECTL,
-    random_name,
-    telepresence_version,
-    run_webserver,
-    create_namespace,
-)
-
-from .rwlock import RWLock
-
-
-REGISTRY = os.environ.get("TELEPRESENCE_REGISTRY", "datawire")
-
-# Some Telepresence configurations take over the machine's whole network
-# stack.  We must avoid concurrency when running those tests.  The network
-# lock gives us a way to do that.  Tests with an isolated network impact
-# (inject-tcp, container) will do a read-acquire on this lock.  Tests with a
-# global network impact (vpn-tcp) will do a write-acquire on this lock.
-network = RWLock()
-
-
-class ResourceIdent(object):
-    """
-    Identify a Kubernetes resource.
-    """
-    def __init__(self, namespace, name):
-        self.namespace = namespace
-        self.name = name
-
-
-def _telepresence(telepresence_args):
-    """
-    Run a probe in a Telepresence execution context.
-    """
-    args = [
-        executable, which("telepresence"),
-        "--logfile", "-",
-    ] + telepresence_args
-    return check_output(
-        args=args,
-        stdin=PIPE,
-        stderr=STDOUT,
-    )
-
-
-
-class _EndToEndTestsMixin(object):
-    """
-    A mixin for ``TestCase`` defining various end-to-end tests for
-    Telepresence.
-    """
-    DESIRED_ENVIRONMENT = {
-        "MYENV": "hello",
-        "EXAMPLE_ENVFROM": "foobar",
-
-        # XXXX
-        # Container method doesn't support multi-line environment variables.
-        # Therefore disable this for all methods or the container tests all
-        # fail...
-        # XXXX
-        # "EX_MULTI_LINE": (
-        #     "first line (no newline before, newline after)\n"
-        #     "second line (newline before and after)\n"
-        # ),
-    }
-
-    def __init__(self, method, operation):
-        """
-        :param _method: The definition of a Telepresence method to use for this
-            test run.
-
-        :param _operation: The definition of a Telepresence operation to use
-            for this test run.
-        """
-        self._method = method
-        self._operation = operation
-
-
-    def test_environment_from_deployment(self):
-        """
-        The Telepresence execution context supplies environment variables with
-        values defined in the Kubernetes Deployment.
-        """
-        probe_environment = self.probe_result["environ"]
-        self.assertEqual(
-            self.DESIRED_ENVIRONMENT,
-            {k: probe_environment.get(k, None) for k in self.DESIRED_ENVIRONMENT},
-            "Probe environment missing some expected items:\n"
-            "Desired: {}\n"
-            "Probed: {}\n".format(self.DESIRED_ENVIRONMENT, probe_environment),
-        )
-
-
-    def test_environment_for_services(self):
-        """
-        The Telepresence execution context supplies environment variables with
-        values locating services configured on the cluster.
-        """
-        probe_environment = self.probe_result["environ"]
-        service_env = self.webserver_name.upper().replace("-", "_")
-        host = probe_environment[service_env + "_SERVICE_HOST"]
-        port = probe_environment[service_env + "_SERVICE_PORT"]
-
-        prefix = service_env + "_PORT_{}_TCP".format(port)
-        desired_environment = {
-            service_env + "_PORT": "tcp://{}:{}".format(host, port),
-            prefix + "_PROTO": "tcp",
-            prefix + "_PORT": port,
-            prefix + "_ADDR": host,
-        }
-
-        self.assertEqual(
-            desired_environment,
-            {k: probe_environment.get(k, None) for k in desired_environment},
-            "Probe environment missing some expected items:\n"
-            "Desired: {}\n"
-            "Probed: {}\n".format(desired_environment, probe_environment),
-        )
-        self.assertEqual(
-            probe_environment[prefix],
-            probe_environment[service_env + "_PORT"],
-        )
-
-
-    def _cleanup_deployment(self, ident):
-        check_call([
-            KUBECTL, "delete",
-            "--namespace", ident.namespace,
-            "--ignore-not-found",
-            "deployment", ident.name,
-        ])
-
-
-
-class _VPNTCPMethod(object):
-    name = "--method=vpn-tcp"
-
-    def lock(self):
-        network.lock_write()
-
-
-    def unlock(self):
-        network.unlock_write()
-
-
-    def telepresence_args(self, probe):
-        return [
-            "--method", "vpn-tcp",
-            "--run", executable, probe,
-        ]
-
-
-
-class _InjectTCPMethod(object):
-    name = "--method=inject-tcp"
-
-    def lock(self):
-        network.lock_read()
-
-
-    def unlock(self):
-        network.unlock_read()
-
-
-    def telepresence_args(self, probe):
-        return [
-            "--method", "inject-tcp",
-            "--run", executable, probe,
-        ]
-
-
-
-class _ContainerMethod(object):
-    name = "--method=container"
-
-    def lock(self):
-        network.lock_read()
-
-
-    def unlock(self):
-        network.unlock_read()
-
-
-    def telepresence_args(self, probe):
-        return [
-            "--method", "container",
-            "--docker-run",
-            "--volume", "{}:/probe.py".format(probe),
-            "python:3-alpine",
-            "python", "/probe.py",
-        ]
-
-
-def create_deployment(deployment_ident, image, environ):
-    def env_arguments(environ):
-        return list(
-            "--env={}={}".format(k, v)
-            for (k, v)
-            in environ.items()
-        )
-    deployment = dumps({
-        "kind": "Deployment",
-        "apiVersion": "extensions/v1beta1",
-        "metadata": {
-            "name": deployment_ident.name,
-            "namespace": deployment_ident.namespace,
-        },
-        "spec": {
-            "replicas": 2,
-            "template": {
-                "metadata": {
-                    "labels": {
-                        "name": deployment_ident.name,
-                        "telepresence-test": deployment_ident.name,
-                    },
-                },
-                "spec": {
-                    "containers": [{
-                        "name": "hello",
-                        "image": image,
-                        "env": list(
-                            {"name": k, "value": v}
-                            for (k, v)
-                            in environ.items()
-                        ),
-                    }],
-                },
-            },
-        },
-    })
-    check_output([KUBECTL, "create", "-f", "-"], input=deployment.encode("utf-8"))
-
-
-
-class _ExistingDeploymentOperation(object):
-    def __init__(self, swap):
-        self.swap = swap
-        if swap:
-            self.name = "--swap-deployment"
-        else:
-            self.name = "--deployment"
-
-
-    def prepare_deployment(self, deployment_ident, environ):
-        if self.swap:
-            image = "openshift/hello-openshift"
-        else:
-            image = "{}/telepresence-k8s:{}".format(
-                REGISTRY,
-                telepresence_version(),
-            )
-        create_deployment(deployment_ident, image, environ)
-
-
-    def telepresence_args(self, deployment_ident):
-        if self.swap:
-            option = "--swap-deployment"
-        else:
-            option = "--deployment"
-        return [
-            "--namespace", deployment_ident.namespace,
-            option, deployment_ident.name,
-        ]
-
-
-
-class _NewDeploymentOperation(object):
-    name = "--new-deployment"
-
-    def prepare_deployment(self, deployment_ident, environ):
-        pass
-
-    def telepresence_args(self, deployment_ident):
-        return [
-            "--namespace", deployment_ident.namespace,
-            "--new-deployment", deployment_ident.name,
-        ]
-
-
-
-def telepresence_tests(method, operation):
-    class Tests(_EndToEndTestsMixin, TestCase):
-        def __init__(self, *args, **kwargs):
-            _EndToEndTestsMixin.__init__(self, method, operation)
-            TestCase.__init__(self, *args, **kwargs)
-    return Tests
-
-
-
-class SwapEndToEndVPNTCPTests(telepresence_tests(
-        _VPNTCPMethod(),
-        _ExistingDeploymentOperation(True),
-)):
-    """
-    Tests for the *vpn-tcp* method using a swapped Deployment.
-    """
-
-
-
-class SwapEndToEndInjectTCPTests(telepresence_tests(
-        _InjectTCPMethod(),
-        _ExistingDeploymentOperation(True),
-)):
-    """
-    Tests for the *inject-tcp* method using a swapped Deployment.
-    """
-
-
-
-class SwapEndToEndContainerTests(telepresence_tests(
-        _ContainerMethod(),
-        _ExistingDeploymentOperation(True),
-)):
-    """
-    Tests for the *container* method using a swapped Deployment.
-    """
-
-
-class ExistingEndToEndVPNTCPTests(telepresence_tests(
-        _VPNTCPMethod(),
-        _ExistingDeploymentOperation(False),
-)):
-    """
-    Tests for the *vpn-tcp* method using an existing Deployment.
-    """
-
-
-
-class ExistingEndToEndInjectTCPTests(telepresence_tests(
-        _InjectTCPMethod(),
-        _ExistingDeploymentOperation(False),
-)):
-    """
-    Tests for the *inject-tcp* method using an existing Deployment.
-    """
-
-
-
-class ExistingEndToEndContainerTests(telepresence_tests(
-        _ContainerMethod(),
-        _ExistingDeploymentOperation(False),
-)):
-    """
-    Tests for the *container* method using an existing Deployment.
-    """
-
-
-class NewEndToEndVPNTCPTests(telepresence_tests(
-        _VPNTCPMethod(),
-        _NewDeploymentOperation(),
-)):
-    """
-    Tests for the *vpn-tcp* method creating a new Deployment.
-    """
-    test_environment_from_deployment = None
-
-
-class NewEndToEndInjectTCPTests(telepresence_tests(
-        _InjectTCPMethod(),
-        _NewDeploymentOperation(),
-)):
-    """
-    Tests for the *inject-tcp* method creating a new Deployment.
-    """
-    test_environment_from_deployment = None
-
-
-
-class NewEndToEndContainerTests(telepresence_tests(
-        _ContainerMethod(),
-        _NewDeploymentOperation(),
-)):
-    """
-    Tests for the *container* method creating a new Deployment.
-    """
-    test_environment_from_deployment = None
